Change type of mass value to `u32` (to simplify testing on GPU later)

+++ crates/gpecs/tests/common/mod.rs
@@ -9,7 +9,7 @@ pub struct Position {
 
 #[derive(Debug, PartialEq, Clone, Copy)]
 pub struct Mass {
-    pub value: u16,
+    pub value: u32,
 }
 
 #[derive(Debug, PartialEq, Clone)]
